requirements.txt
requests>=2.31.0
supabase>=2.0.0
python-dateutil>=2.8.2
python-dotenv
streamlit>=1.31.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
anthropic>=0.18.1
Pillow>=10.3.0
html5lib>=1.1
plotly>=5.18.0
pandas>=2.0.0

